feat: 用户ID类型改为Long并优化请求限流逻辑
将用户ID从Integer改为Long类型，涉及多个实体类和工具类
重构请求限流逻辑，使用Lua脚本实现原子操作，支持用户ID和IP两种限流维度

+++ src/main/java/cloud/bytepulse/bp/common/util/AuthUtils.java
@@ -32,7 +32,7 @@ public class AuthUtils {
         if ("anonymousUser".equals(authentication.getPrincipal())) {
             LoginUser loginUser = new LoginUser();
             LoginUserInfo loginUserInfo = new LoginUserInfo();
-            loginUserInfo.setUserId(0);
+            loginUserInfo.setUserId(0L);
             loginUserInfo.setUsername("anonymousUser");
             loginUserInfo.setNickname("anonymousUser");
             loginUser.setLoginUserInfo(loginUserInfo);
@@ -55,10 +55,10 @@ public class AuthUtils {
     /**
      * 获取当前用户id
      */
-    public static Integer getUserId() {
+    public static Long getUserId() {
         LoginUserInfo user = getUser();
         if (user == null) {
-            return 0;
+            return 0L;
         }
         return user.getUserId();
     }
